# openapi.json
{"openapi": "3.1.0", "info": {"title": "Lock Detection API", "description": "\u667a\u80fd\u9501\u68c0\u6d4b\u7cfb\u7edf - \u57fa\u4e8eYOLOv8\u7684\u9501\u72b6\u6001\u8bc6\u522bAPI\n                    \u652f\u6301\u9489\u9489\u673a\u5668\u4eba\u56de\u8c03\u3001\u56fe\u7247\u68c0\u6d4b\u3001\u6a21\u578b\u8bad\u7ec3\u7b49\u529f\u80fd", "version": "2025.09.04"}, "paths": {"/healthcheck": {"get": {"summary": "Perform Healthcheck", "description": "\u5b83\u53d1\u9001\u4e00\u4e2aGET\u8bf7\u6c42\u5230\u8be5\u8def\u7531\uff0c\u5e76\u5e0c\u671b\u5f97\u5230\u4e00\u4e2a\"200\"\u54cd\u5e94\u4ee3\u7801\u3002\n\u672a\u80fd\u8fd4\u56de200\u54cd\u5e94\u4ee3\u7801\u5c06\u4f7fGitHub Actions\u56de\u6eda\u5230\u9879\u76ee\u5904\u4e8e\"\u5de5\u4f5c\u72b6\u6001\"\u7684\u6700\u540e\u4e00\u4e2a\u7248\u672c\u3002\n\u5b83\u4f5c\u4e3a\u6700\u540e\u4e00\u9053\u9632\u7ebf\uff0c\u4ee5\u9632\u53d1\u751f\u95ee\u9898\u3002\n\u6b64\u5916\uff0c\u5b83\u8fd8\u4ee5JSON\u683c\u5f0f\u8fd4\u56de\u54cd\u5e94\uff0c\u5f62\u5f0f\u4e3a\uff1a\n{\n    'healthcheck': '\u4e00\u5207\u6b63\u5e38\uff01'\n}", "operationId": "perform_healthcheck_healthcheck_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/img_object_detection_to_json": {"post": {"summary": "Img Object Detection To Json", "description": "\u4ece\u56fe\u50cf\u4e2d\u8fdb\u884c\u5bf9\u8c61\u68c0\u6d4b\u3002\n\n\u53c2\u6570:\n    file (bytes): \u4ee5\u5b57\u8282\u683c\u5f0f\u7684\u56fe\u50cf\u6587\u4ef6\u3002\n\u8fd4\u56de:\n    dict: \u5305\u542b\u5bf9\u8c61\u68c0\u6d4b\u7ed3\u679c\u7684JSON\u683c\u5f0f\u3002", "operationId": "img_object_detection_to_json_img_object_detection_to_json_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_object_detection_to_json_img_object_detection_to_json_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/img_object_detection_to_img": {"post": {"summary": "Img Object Detection To Img", "description": "\u4ece\u56fe\u50cf\u4e2d\u8fdb\u884c\u5bf9\u8c61\u68c0\u6d4b\u5e76\u5728\u56fe\u50cf\u4e0a\u7ed8\u5236\u8fb9\u754c\u6846\n\n\u53c2\u6570:\n    file (bytes): \u4ee5\u5b57\u8282\u683c\u5f0f\u7684\u56fe\u50cf\u6587\u4ef6\u3002\n\u8fd4\u56de:\n    Image: \u5e26\u6709\u8fb9\u754c\u6846\u6ce8\u91ca\u7684\u5b57\u8282\u683c\u5f0f\u56fe\u50cf\u3002", "operationId": "img_object_detection_to_img_img_object_detection_to_img_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_object_detection_to_img_img_object_detection_to_img_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/lock/detect": {"post": {"summary": "Detect Locks", "description": "\u68c0\u6d4b\u56fe\u7247\u4e2d\u7684\u9501\u72b6\u6001\n\n\u53c2\u6570:\n    file (bytes): \u56fe\u7247\u6587\u4ef6\n    user_id (str): \u7528\u6237ID\uff0c\u53ef\u9009\n\n\u8fd4\u56de:\n    DetectionResponse: \u68c0\u6d4b\u7ed3\u679c", "operationId": "detect_locks_api_v1_lock_detect_post", "parameters": [{"name": "user_id", "in": "query", "required": false, "schema": {"type": "string", "default": "", "title": "User Id"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_detect_locks_api_v1_lock_detect_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetectionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/dingtalk/configure": {"post": {"summary": "Configure Dingtalk", "description": "\u914d\u7f6e\u9489\u9489\u673a\u5668\u4eba\n\n\u53c2\u6570:\n    config (DingTalkConfig): \u9489\u9489\u914d\u7f6e\n\n\u8fd4\u56de:\n    DetectionResponse: \u914d\u7f6e\u7ed3\u679c", "operationId": "configure_dingtalk_api_v1_dingtalk_configure_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DingTalkConfig"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetectionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/dingtalk/webhook": {"post": {"summary": "Dingtalk Webhook", "description": "\u9489\u9489\u56de\u8c03\u63a5\u53e3\n\n\u53c2\u6570:\n    request (Request): HTTP\u8bf7\u6c42\n\n\u8fd4\u56de:\n    JSONResponse: \u5904\u7406\u7ed3\u679c", "operationId": "dingtalk_webhook_api_v1_dingtalk_webhook_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/stats": {"get": {"summary": "Get Statistics", "description": "\u83b7\u53d6\u7cfb\u7edf\u7edf\u8ba1\u4fe1\u606f\n\n\u8fd4\u56de:\n    DetectionResponse: \u7edf\u8ba1\u4fe1\u606f", "operationId": "get_statistics_api_v1_stats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetectionResponse"}}}}}}}, "/api/v1/history": {"get": {"summary": "Get Detection History", "description": "\u83b7\u53d6\u68c0\u6d4b\u5386\u53f2\n\n\u53c2\u6570:\n    limit (int): \u9650\u5236\u6570\u91cf\n    offset (int): \u504f\u79fb\u91cf\n\n\u8fd4\u56de:\n    DetectionResponse: \u68c0\u6d4b\u5386\u53f2", "operationId": "get_detection_history_api_v1_history_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 10, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetectionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/health": {"get": {"summary": "Health Check", "description": "\u5065\u5eb7\u68c0\u67e5\u63a5\u53e3\n\n\u8fd4\u56de:\n    DetectionResponse: \u5065\u5eb7\u72b6\u6001", "operationId": "health_check_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetectionResponse"}}}}}}}}, "components": {"schemas": {"Body_detect_locks_api_v1_lock_detect_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_detect_locks_api_v1_lock_detect_post"}, "Body_img_object_detection_to_img_img_object_detection_to_img_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_img_object_detection_to_img_img_object_detection_to_img_post"}, "Body_img_object_detection_to_json_img_object_detection_to_json_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_img_object_detection_to_json_img_object_detection_to_json_post"}, "DetectionResponse": {"properties": {"success": {"type": "boolean", "title": "Success"}, "message": {"type": "string", "title": "Message"}, "data": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Data"}}, "type": "object", "required": ["success", "message"], "title": "DetectionResponse"}, "DingTalkConfig": {"properties": {"app_key": {"type": "string", "title": "App Key"}, "app_secret": {"type": "string", "title": "App Secret"}, "webhook_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Webhook Url", "default": ""}}, "type": "object", "required": ["app_key", "app_secret"], "title": "DingTalkConfig"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}